SDWEAK v1.4

+++ uninstall.sh
@@ -56,7 +56,16 @@ red_msg "Uninstalling..."
 sudo steamos-readonly disable
 sudo systemctl disable sshd
 # Pacman
-sudo pacman -Sy
+sudo rm -rf /home/.steamos/offload/var/cache/pacman/pkg/{*,.*}
+sudo rm -rf /etc/pacman.d/gnupg
+sudo pacman-key --init
+sudo pacman-key --populate
+if ! sudo pacman -Sy; then
+    exit 1
+fi
+if ! sudo pacman -S --noconfirm sed; then
+    exit 1
+fi
 
 # Yet-tweak
 sudo rm -f /etc/tmpfiles.d/mglru.conf
@@ -67,27 +76,13 @@ sudo grub-mkconfig -o /boot/efi/EFI/steamos/grub.cfg
 sudo systemctl unmask steamos-cfs-debugfs-tunings.service
 sudo systemctl unmask gpu-trace.service
 sudo systemctl unmask steamos-log-submitter.service
-sudo systemctl unmask steamos-devkit-service.service
 sudo systemctl unmask cups.service
-sudo systemctl unmask firewalld.service
 sudo systemctl unmask gamemoded.service
 sudo systemctl unmask avahi-daemon.service
 sudo systemctl unmask avahi-daemon.socket
-
-sudo systemctl start steamos-cfs-debugfs-tunings.service
-sudo systemctl start gpu-trace.service
-sudo systemctl start steamos-log-submitter.service
-sudo systemctl start steamos-devkit-service.service
-sudo systemctl start cups.service
-sudo systemctl start firewalld.service
-sudo systemctl start gamemoded.service
-sudo systemctl start avahi-daemon.service
-sudo systemctl start avahi-daemon.socket
-cp -f $HOME/install_backup/50-coredump.conf /usr/lib/sysctl.d/50-coredump.conf
-cp -f $HOME/install_backup/21-steamos-panic-sysctls.conf /usr/lib/sysctl.d/21-steamos-panic-sysctls.conf
-cp -f $HOME/install_backup/20-panic-sysctls.conf /usr/lib/sysctl.d/20-panic-sysctls.conf
-cp -f $HOME/install_backup/20-sched.conf /usr/lib/sysctl.d/20-sched.conf
-cp -f $HOME/install_backup/60-crash-hook.conf /usr/lib/sysctl.d/60-crash-hook.conf
+sudo cp -f $HOME/install_backup/50-coredump.conf /usr/lib/sysctl.d/50-coredump.conf
+sudo cp -f $HOME/install_backup/20-sched.conf /usr/lib/sysctl.d/20-sched.conf
+sudo cp -f $HOME/install_backup/60-crash-hook.conf /usr/lib/sysctl.d/60-crash-hook.conf
 
 # Daemon uninstall
 sudo systemctl disable ananicy-cpp
@@ -104,11 +99,11 @@ sudo rm -f /usr/lib/systemd/zram-generator.conf
 sudo pacman -Rdd --noconfirm holo-zram-swap zram-generator
 sudo pacman -S --noconfirm --needed holo-zram-swap zram-generator
 sudo systemctl restart systemd-zram-setup@zram0
-#THP
+# THP
 sudo rm -f /usr/lib/tmpfiles.d/thp-shrinker.conf
 sudo rm -f /usr/lib/tmpfiles.d/thp.conf
 
-sudo sed -i "s/ENABLE_GAMESCOPE_WSI=0/ENABLE_GAMESCOPE_WSI=1/g" /usr/{bin/gamescope-session,lib/steamos/gamescope-session/gamescope-session}
+sudo sed -i "s/ENABLE_GAMESCOPE_WSI=0/ENABLE_GAMESCOPE_WSI=1/g" /usr/{bin/gamescope-session,lib/steamos/gamescope-session}
 sudo pacman -S --noconfirm --needed vulkan-radeon lib32-vulkan-radeon
 
 sudo sed -z -i "s/58, 59,\n        60, 61, 62, 63, 64, 65, 66, 67, 68, 69,\n        70/58, 59,\n        60/g" /usr/share/gamescope/scripts/00-gamescope/displays/valve.steamdeck.lcd.lua
@@ -125,23 +120,18 @@ sudo systemctl disable --now energy.timer
 sudo rm -f /etc/systemd/system/energy.service
 sudo rm -f /etc/systemd/system/energy.timer
 
+if { [ "$steamos_version" = "3.7" ] || [ "$steamos_version" = "3.8" ]; }; then
+    sudo pacman -S --noconfirm linux-neptune-611 linux-neptune-611-headers
+fi
 
-if [ $steamos_version = 3.7 ]
-then
-    sudo pacman -S --noconfirm linux-neptune-611
-fi
-if [ $steamos_version = 3.6 ]
-then
-    sudo pacman -S --noconfirm linux-neptune-65
-    sudo pacman -R --noconfirm linux-neptune-611
-    sudo pacman -R --noconfirm linux-neptune-611-headers
-    sudo sed -i "s/3.7/3.6/g" /etc/pacman.conf
-    sudo sed -i "s/main/3.6/g" /etc/pacman.conf
-    sudo pacman -Sy --noconfirm ell readline iwd networkmanager steamos-networking-tools steamos-manager iptables linux-api-headers jupiter-firewall linux-firmware-neptune linux-firmware-neptune-whence &>/dev/null
-fi
-if [ $steamos_version = 3.8 ]
-then
-    sudo pacman -S --noconfirm linux-neptune-611
-fi
+params=("amdgpu.moverate=128" "amdgpu.mes=1" "amdgpu.cwsr_enable=0" "amdgpu.umsch_mm=1" "amdgpu.uni_mes=1")
+
+for param in "${params[@]}"; do
+    if grep -q "$param" /etc/default/grub &>/dev/null; then
+        sudo sed -i "s/\b$param\b//g" /etc/default/grub &>/dev/null
+    else
+        echo 1 > /dev/null
+    fi
+done
 sudo grub-mkconfig -o /boot/efi/EFI/steamos/grub.cfg &>/dev/null
 sudo systemctl daemon-reload
